extract resolveQuoteDefaults and computeSlippageBounds to base class
